[pylint] Rename constant-comparison to comparison-of-constant
The actual name
