+++ resources/test/fixtures/pylint/comparison_of_constant.py
