+++ src/rules/pylint/rules/mod.rs
@@ -1,6 +1,6 @@
 pub use await_outside_async::{await_outside_async, AwaitOutsideAsync};
+pub use comparison_of_constant::{comparison_of_constant, ComparisonOfConstant};
 pub use consider_using_sys_exit::{consider_using_sys_exit, ConsiderUsingSysExit};
-pub use constant_comparison::{constant_comparison, ConstantComparison};
 pub use global_variable_not_assigned::GlobalVariableNotAssigned;
 pub use invalid_all_format::{invalid_all_format, InvalidAllFormat};
 pub use invalid_all_object::{invalid_all_object, InvalidAllObject};
@@ -21,8 +21,8 @@ pub use useless_else_on_loop::{useless_else_on_loop, UselessElseOnLoop};
 pub use useless_import_alias::{useless_import_alias, UselessImportAlias};
 
 mod await_outside_async;
+mod comparison_of_constant;
 mod consider_using_sys_exit;
-mod constant_comparison;
 mod global_variable_not_assigned;
 mod invalid_all_format;
 mod invalid_all_object;
